Add dynamic theme icon
Adds dynamic theme icon that changes with user's theme colors.
The change was easy and has been tested on API levels 24, 26, 28, 31, 34.
The old icon in unchanged but a new foreground icon is added which is
vector drawable. It replaces the default foreground icon that was
present but wasn't being used. The new foreground icon is then referenced
as the monochrome icon.