Change SwipeRefreshLayout's progress view position
There are two position for the progress view - start and end/loading
positions. Start position is the easy one - it is always just outside
the visible area, same as by default with SwipeRefreshLayout.

For the end position I had two options. The first one is to always keep the
distance between start and end position same as default. However, in
theory status bar's height can be greater than that default distance, so
status bar's content would overlap with progress view in its loading
state. So I chose the second option - keep the position relative to
screen's edges constant. This way end position is always the same
distance from the bottom edge of the status bar. This means that with
transparent status bar the distance is greater than the default, however
it guarantees that progress view won't intersect with status bar in its
loading state.