Disable refresh while typing the search query
There is no logical reason to refresh while you are typing. However, the
refresh would have revealed an invalid cache and would start loading
the results from the previous search.

You could reproduce this with:
1. Search something
2. Go to the main screen
3. Go to the search
4. Refresh by pulling from the top

Now the refresh gesture is disabled while you are in the search and no
results have been loaded.