Rework comments recycler paddings
Set recycler padding equal to status bar.

The main difficulty was to preserve bottom sheet sliding animation.
Previously it was achieved by changing padding of header inside
recycler. Now that padding is moved to the recycler itself and changing
it while dragging the bottom sheet caused severe lag or other problems.
So insted the padding doesn't change and instead recycler's `translationY`
is changed.