Use onDestroyView fragment methods for cleanup
This addresses two problems.

The first one is a rare crash in CommentsFragment that can happen if
fragment gets destroyed after executing onCreate but before
onViewCreated. In this case webView is null which causes
NullPointerException.

The second one is a memory leak that can happen if fragment's view is
destroyed, but fragment itself is not. In this case requests will still
hold references to the layout in their listeners but the layout is
supposed to be free for garbage collection at this point. Note that
technically this doesn't change anything because the fragment itself
still holds references to the views. Still, this code is less wrong
than the old version.