Always add translucent nav bar scrim before API 29
Doing it the old way because the new method would enable edge-to-edge
in compose activity where we don't use it. And also because I don't like
the new method since it is inconvenient to use and behaves in unintuitive
(though documented) ways on some API levels.

Fixes #79, at least until Google changes something again.