Feat: Add option to hide comments count in list view
This comment adds a feature to hide the number of comments of a story in the list view. It replaces the number of comments with a dot if there is any, or nothing if there is no comment.

This allows a more zen approach to scrolling, helping appeal to popularity to which our attention is drawn to before reading a story.

Option is disabled by default, and does not show-up during initial setup.