Refactor comments, imports and some code pieces
I know that might come of as rude as it wasn't requested but it really
shouldn't. I just wanted to read all parts of the app and stumbled upon
a couple of things I wanted to fix.

If you have any suggestions I am more than willing to revert some
changes here.