Add a changelog button to the about page
Previously the changelog was only displayed after and update, however it
is sometimes quite handy to view the changelog to see what changed in
the recent versions. Moreover I sometimes don't have the time to check
the changelog when the update is released but want to read it a little
bit later.