Display the debug indicator in the about section of the app
While deploying a custom version of the app to my phone I wanted to be
sure that the deployed app really is the release version.

Therefore I added this warning in the about section if the app is not
a release version. While most users hopefully should never see this, it
should make it easier for devs to troubleshoot problems.

The commit also adds padding to the last button on the about page so
that it doesn't hug the bottom that much.